refactor : type 경로 수정
